Move key encoding to key package
