+++ database/table_test.go
@@ -8,9 +8,9 @@ import (
 
 	"github.com/genjidb/genji/database"
 	"github.com/genjidb/genji/document"
-	"github.com/genjidb/genji/document/encoding"
 	"github.com/genjidb/genji/document/encoding/msgpack"
 	"github.com/genjidb/genji/engine/memoryengine"
+	"github.com/genjidb/genji/key"
 	"github.com/stretchr/testify/require"
 )
 
@@ -177,10 +177,10 @@ func TestTableInsert(t *testing.T) {
 
 		key2 := insertDoc(db)
 
-		a, err := encoding.DecodeInt64(key1)
+		a, err := key.DecodeInt64(key1)
 		require.NoError(t, err)
 
-		b, err := encoding.DecodeInt64(key2)
+		b, err := key.DecodeInt64(key2)
 		require.NoError(t, err)
 
 		require.Equal(t, a+1, b)
@@ -197,7 +197,7 @@ func TestTableInsert(t *testing.T) {
 		require.NoError(t, err)
 
 		manualInsert := func(id int64) {
-			docid := encoding.EncodeInt64(id)
+			docid := key.EncodeInt64(id)
 			v, err := msgpack.EncodeDocument(newDocument())
 			require.NoError(t, err)
 			err = tb.Store.Put(docid, v)
@@ -210,7 +210,7 @@ func TestTableInsert(t *testing.T) {
 		manualInsert(2)
 
 		expectDocid := func(want int64, got []byte) {
-			newDocid, err := encoding.DecodeInt64(got)
+			newDocid, err := key.DecodeInt64(got)
 			require.NoError(t, err)
 			require.Equal(t, want, newDocid)
 		}
@@ -244,16 +244,16 @@ func TestTableInsert(t *testing.T) {
 		require.NoError(t, err)
 
 		// insert
-		key, err := tb.Insert(doc)
+		k, err := tb.Insert(doc)
 		require.NoError(t, err)
-		require.Equal(t, encoding.EncodeInt64(10), key)
+		require.Equal(t, key.EncodeInt64(10), k)
 
 		// make sure the document is fetchable using the returned key
-		_, err = tb.GetDocument(key)
+		_, err = tb.GetDocument(k)
 		require.NoError(t, err)
 
 		// insert again
-		key, err = tb.Insert(doc)
+		k, err = tb.Insert(doc)
 		require.Equal(t, database.ErrDuplicateDocument, err)
 	})
 
@@ -348,7 +348,7 @@ func TestTableInsert(t *testing.T) {
 		require.NoError(t, err)
 
 		var count int
-		err = idx.AscendGreaterOrEqual(nil, func(val document.Value, k []byte) error {
+		err = idx.AscendGreaterOrEqual(nil, func(val, k []byte) error {
 			switch count {
 			case 0:
 				// key2, which doesn't countain the field must appear first in the next,
